extracting types into its own files where appropriate, excluding from test coverage classes that don't have to be tested
