+++ GitHubWebUserFinder/App_Start/FilterConfig.cs
@@ -1,8 +1,10 @@
-﻿using System.Web;
+﻿using System.Diagnostics.CodeAnalysis;
+using System.Web;
 using System.Web.Mvc;
 
 namespace GitHubWebUserFinder
 {
+	[ExcludeFromCodeCoverage]
 	public class FilterConfig
 	{
 		public static void RegisterGlobalFilters(GlobalFilterCollection filters)
