+++ GitHubWebUserFinder/App_Start/RouteConfig.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
@@ -7,6 +8,7 @@ using System.Web.Routing;
 
 namespace GitHubWebUserFinder
 {
+	[ExcludeFromCodeCoverage]
 	public class RouteConfig
 	{
 		public static void RegisterRoutes(RouteCollection routes)
