+++ GitHubWebUserFinder/Connectors/IAppHttpClient.cs
@@ -1,0 +1,10 @@
+﻿using System.Net.Http;
+using System.Threading.Tasks;
+
+namespace GitHubWebUserFinder.Connectors
+{
+	public interface IAppHttpClient
+	{
+		Task<HttpResponseMessage> Get(string url);
+	}
+}
