+++ GitHubWebUserFinder/Connectors/IGitHubConnector.cs
@@ -1,0 +1,10 @@
+﻿using GitHubWebUserFinder.Models;
+using System.Threading.Tasks;
+
+namespace GitHubWebUserFinder.Connectors
+{
+	public interface IGitHubConnector
+	{
+		Task<GitHubUser> FindUser(string criteria);
+	}
+}
