+++ GitHubWebUserFinder/GitHubWebUserFinder.csproj
@@ -154,7 +154,10 @@
     <Compile Include="App_Start\FilterConfig.cs" />
     <Compile Include="App_Start\RouteConfig.cs" />
     <Compile Include="App_Start\UnityConfig.cs" />
+    <Compile Include="Connectors\AppHttpClient.cs" />
     <Compile Include="Connectors\GitHubConnector.cs" />
+    <Compile Include="Connectors\IAppHttpClient.cs" />
+    <Compile Include="Connectors\IGitHubConnector.cs" />
     <Compile Include="Controllers\ResultController.cs" />
     <Compile Include="Controllers\HomeController.cs" />
     <Compile Include="Global.asax.cs">
@@ -165,6 +168,7 @@
     <Compile Include="Models\UserSearch.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\GitHubSearchService.cs" />
+    <Compile Include="Services\IGitHubSearchService.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="Content\bootstrap-theme.css" />
