+++ GitHubWebUserFinder/Global.asax.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
@@ -8,6 +9,7 @@ using System.Web.Routing;
 
 namespace GitHubWebUserFinder
 {
+	[ExcludeFromCodeCoverage]
 	public class MvcApplication : System.Web.HttpApplication
 	{
 		protected void Application_Start()
