+++ GitHubWebUserFinder/Services/IGitHubSearchService.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+using GitHubWebUserFinder.Models;
+
+namespace GitHubWebUserFinder.Services
+{
+	public interface IGitHubSearchService
+	{
+		Task<GitHubUser> FindUser(string name);
+	}
+}
